Project Structure
So I ran across an issue, if we all try to push our own builds of ionic, there will be a butt ton of conflicts (due to project build files and crap). So we will be running a local project and divide out screens for eachother. I will discuss this in our meeting. Right now there should be 3 folders, home page has 3 files in it.